rewrite KSink::get_server_model return uint32_t

+++ include/KHttp3Sink.h
@@ -180,7 +180,7 @@ public:
 			lsquic_stream_flush(st);
 		}
 	}
-	uint8_t get_server_model() override
+	uint32_t get_server_model() override
 	{
 		return 0;
 	}
