+++ include/KTcpServerSink.h
@@ -12,7 +12,7 @@ public:
 	{
 
 	}
-	uint8_t get_server_model() override
+	uint32_t get_server_model() override
 	{
 		return get_server()->flags;
 	}
